+++ src/main/java/ru/itmo/sortvis/GraphWalkerListener.java
@@ -4,6 +4,6 @@ public interface GraphWalkerListener {
     void nodeIn(String node);
     void nodeOut(String node);
 
-    void edgeForward(int i, int j);
-    void edgeBack(int i, int j);
+    void edgeForward(long i, long j);
+    void edgeBack(long i, long j);
 }
